pyproject.toml
[build-system]
requires = ["maturin>=1.7,<2.0"]
build-backend = "maturin"

[project]
name = "wse-server"
version = "1.0.0"
description = "Real-time WebSocket engine for Python. Up to 1.1M msg/s. End-to-end encrypted. Rust-accelerated."
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "WSE Contributors" },
]
keywords = [
    "websocket",
    "engine",
    "real-time",
    "pubsub",
    "fastapi",
    "encrypted",
    "rust",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Programming Language :: Rust",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: System :: Networking",
    "Typing :: Typed",
]

dependencies = [
    "fastapi>=0.115",
    "starlette>=0.40",
    "orjson>=3.10",
    "redis>=5.0",
    "cryptography>=43.0",
    "pyjwt[crypto]>=2.9",
]

[project.optional-dependencies]
msgpack = [
    "msgpack>=1.0",
]
prometheus = [
    "prometheus-client>=0.21",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "httpx>=0.27",
    "uvicorn>=0.30",
    "ruff>=0.8",
    "mypy>=1.13",
    "maturin>=1.7",
]

[project.urls]
Homepage = "https://github.com/silvermpx/wse"
Documentation = "https://github.com/silvermpx/wse/tree/main/docs"
Repository = "https://github.com/silvermpx/wse"
Issues = "https://github.com/silvermpx/wse/issues"
Changelog = "https://github.com/silvermpx/wse/releases"

[tool.maturin]
features = ["pyo3/extension-module", "pyo3/abi3-py312"]
module-name = "wse_server._wse_accel"
manifest-path = "rust/Cargo.toml"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "TCH"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"wse_server/examples/*" = ["F401", "F821", "N806"]
"wse_server/reliability/circuit_breaker.py" = ["SIM102"]
"wse_server/connection/connection.py" = ["SIM102"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
